Remove plc_docker_common.[ch] and move the code to docker api file.
Those file exist because long ago we have two versions of docker REST API access code, but now we have just one version (i.e. using libucrl).
