+++ src/plc_backend_api.c
@@ -6,6 +6,7 @@
  */
 
 #include "plc_backend_api.h"
+#include "plc_docker_api.h"
 #include "common/comm_utils.h"
 
 static PLC_FunctionEntriesData *CurrentBackend;
