+++ src/plc_docker_api.h
@@ -10,7 +10,6 @@
 #define PLC_DOCKER_API_H
 
 #include "plc_configuration.h"
-#include "plc_docker_common.h"
 
 #define CURL_BUFFER_SIZE 8192
 
@@ -42,5 +41,7 @@ int plc_docker_delete_container(const char *name);
 int plc_docker_list_container(char **result) __attribute__((warn_unused_result));
 
 int plc_docker_get_container_state(const char *name, char **result) __attribute__((warn_unused_result));
+
+int docker_inspect_string(char *buf, char **element, plcInspectionMode type);
 
 #endif /* PLC_DOCKER_API_H */
